notifiers/image-tagger/pyproject.toml
[tool.poetry]
authors = ["Brandon Bjelland <brandon@atscale.run>"]
description = ""
name = "image-tagger"
packages = [{include = "src"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
fastapi = "^0.89.1"
google-cloud-deploy = "^1.6.1"
python = "^3.10"
# TODO(bjb): bump tags after 0.2.0 release
python-clouddeploy-notifier = {git = "https://github.com/computeclub/gcp-cloud-deploy-notifiers.git", subdirectory = "lib/python-notifier", tag = "v0.1.0"}
# python-clouddeploy-notifier = {path = "../../lib/python-notifier", develop = true}
google-cloud-artifact-registry = "^1.7.0"
uvicorn = "^0.20.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
